requirements.txt
Django==2.2.6
django_autoslug==1.9.6
django_mptt==0.10.0
djangorestframework==3.10.3
